script-addon.sh: recursively pass files to functions, not dirs
The addon.d helper functions don't like being passed directories, so we
just use find and pass the them the files ourselves.

We could have implemented the helper functions, but there's too much of
a mess going on in the addon.d world. That will probably cause more problems
that it will solve.