npem: install microG APKs as user app
The init script was used to do it before, but init.d has been all but
abandoned by ROMs universally, so effectively the only situation in which
the init script runs is for Magisk.

This change will be very helpful for non-Magisk users, as they will no
longer have to extract and install the APK by hand.

Also deduplicate code from init script and make it just call npem.